style: separate mocks for log manager tests
