+++ packages/log_manager/test/custom_test_platform_dispatcher.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use, no-empty-block
 
 import 'dart:isolate';
 import 'dart:typed_data';
