+++ packages/log_manager/test/log_manager_test.dart
@@ -1,32 +1,9 @@
-import 'dart:async';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:log_manager/log_manager.dart';
 import 'package:mocktail/mocktail.dart';
 
-class MockLogManager extends LogManager with Mock {
-  final StreamController<BaseLogMessage> _streamController =
-      StreamController<BaseLogMessage>();
-
-  @override
-  Stream<BaseLogMessage> get logStream => _streamController.stream;
-
-  void addLogMessage(BaseLogMessage message) {
-    _streamController.add(message);
-  }
-
-  @override
-  Future<void> close() async {
-    await _streamController.close();
-  }
-
-  @override
-  Future<void> logFlutterError(FlutterErrorDetails details) async {
-    // Override only if needed for specific behavior
-    return super.logFlutterError(details);
-  }
-}
+import 'mocks/mock_log_manager.dart';
 
 void main() {
   group('LogManager', () {
