+++ packages/log_manager/test/mocks/mock_build_mode.dart
@@ -1,0 +1,4 @@
+import 'package:log_manager/src/logger_build_mode.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockBuildMode extends Mock implements LoggerBuildMode {}
