+++ packages/log_manager/test/mocks/mock_console_logger.dart
@@ -1,0 +1,11 @@
+import 'package:log_manager/src/output-events/dev_console_output.dart';
+import 'package:mocktail/mocktail.dart';
+
+final class MockConsoleLogger extends Mock implements CustomConsoleLogger {
+  final List<String> loggedMessages = <String>[];
+
+  @override
+  void log(String message) {
+    loggedMessages.add(message);
+  }
+}
