+++ packages/log_manager/test/mocks/mock_logger.dart
@@ -1,0 +1,8 @@
+import 'package:logger/logger.dart';
+import 'package:mocktail/mocktail.dart';
+
+final class MockLogger extends Mock implements Logger {
+  @override
+  Future<void> close() async =>
+      super.noSuchMethod(Invocation.method(#close, <Object?>[]));
+}
