+++ packages/log_manager/test/mocks/mock_logger_wrapper.dart
@@ -1,0 +1,4 @@
+import 'package:log_manager/src/logger_output_wrapper.dart';
+import 'package:mocktail/mocktail.dart';
+
+final class MockLoggerWrapper extends Mock implements LoggerOutputWrapper {}
