+++ packages/log_manager/test/output-events/dev_console_output_test.dart
@@ -1,19 +1,11 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:log_manager/log_manager.dart';
 
-class MockConsoleLogger implements CustomConsoleLogger {
-  final List<String> loggedMessages = <String>[];
-
-  @override
-  void log(String message) {
-    loggedMessages.add(message);
-  }
-}
+import '../mocks/mock_console_logger.dart';
 
 void main() {
   group('DevConsoleOutput', () {
     test('should log messages using the provided CustomConsoleLogger', () {
-      // Arrange
       final MockConsoleLogger mockLogger = MockConsoleLogger();
       final DevConsoleOutput devConsoleOutput = DevConsoleOutput(mockLogger);
       const String testMessage = 'Test log message';
@@ -23,14 +15,12 @@ void main() {
         <String>[testMessage],
       );
 
-      // Act
       devConsoleOutput.output(outputEvent);
 
       expect(mockLogger.loggedMessages, contains(testMessage));
     });
 
     test('should handle multiple lines of log output', () {
-      // Arrange
       final MockConsoleLogger mockLogger = MockConsoleLogger();
       final DevConsoleOutput devConsoleOutput = DevConsoleOutput(mockLogger);
 
@@ -40,10 +30,8 @@ void main() {
         <String>['Line 1', 'Line 2', 'Line 3'],
       );
 
-      // Act
       devConsoleOutput.output(outputEvent);
 
-      // Assert
       expect(mockLogger.loggedMessages, contains('Line 1\nLine 2\nLine 3'));
     });
   });
